requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.8
numpy>=1.24.0
opencv-python>=4.8.0.76
Pillow>=9.5.0
PyYAML>=6.0.1
tqdm>=4.65.0
